% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Births}
\alias{Births}
\title{Births dataset}
\format{
A data frame with N rows and K variables:
\describe{
  \item{date}{Date}
  \item{births}{Observed value}
}
}
\usage{
Births
}
\description{
Dataset containing births data from 01-01-1968 to 2024-12-31.
}
\keyword{datasets}
