% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_estimation}
\alias{.extended_airline_estimation}
\title{Estimate parameters of an Extended Airline RegARIMA model}
\usage{
.extended_airline_estimation(
  jregarima,
  jspec,
  eps = 1e-09,
  deps = 1e-09,
  exact_hessian = FALSE
)
}
\arguments{
\item{jregarima}{A Java RegArimaModel object, for instance created using
\code{\link{.extended_airline_regarima}}.}

\item{jspec}{A Java ExtendedAirlineSpec object, for instance created using
\code{\link{.extended_airline_spec}}.}

\item{eps}{Numeric scalar specifying the convergence tolerance for the
optimization algorithm.}

\item{deps}{Numeric scalar. Step in the computation of the numerical derivatives,
used in the optimisation routine. Default:1e-4.}

\item{exact_hessian}{Logical. If \code{TRUE}, computes the exact Hessian matrix
at the optimum for calculating standard errors. If \code{FALSE} (default),
uses a numerical approximation.}
}
\value{
A list object containing detailed estimation results.
}
\description{
This internal function performs maximum likelihood estimation of an Extended
Airline model that has been previously specified and initialized.
}
\keyword{internal}
