% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_loglevel}
\alias{.extended_airline_loglevel}
\title{Log-Level Test for Extended Airline Models}
\usage{
.extended_airline_loglevel(jregarima, jspec, precision = 1e-05, deps = 1e-04)
}
\arguments{
\item{jregarima}{Java object containing the already estimated RegARIMA model.}

\item{jspec}{Java object containing the Extended Airline model specifications.}

\item{precision}{Numeric value specifying the tolerance for convergence
of optimization algorithms. Default: 1e-5.}

\item{deps}{Numeric scalar. Step in the computation of the numerical derivatives,
used in the optimisation routine. Default:1e-4.}
}
\value{
An object containing the log-level test results.
}
\description{
This internal function performs a statistical test to determine whether data
should be transformed to logarithmic scale or kept in level (original scale)
}
\examples{
jspec     <- .extended_airline_spec(c(7))
jregarima <-  rjd3highfreq::.extended_airline_regarima(y = Births$births, jspec=jspec)
aiccs     <-  rjd3highfreq::.extended_airline_loglevel(jregarima = jregarima,   jspec = jspec) # {ll.getAICcLevel(), ll.getAICcLog()}, in Java code
if(aiccs[1] <= aiccs[2]) {
    print("AICC is higher in the model without log transformation")
}  else {
    print("AICC is higher in the model with log transformation")
}
}
