% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_outliers}
\alias{.extended_airline_outliers}
\title{Title}
\usage{
.extended_airline_outliers(
  jregarima,
  jspec,
  types = c("ao"),
  start = 0,
  end = 0,
  critical_value = 0,
  max_outliers = 30,
  max_round = 30,
  precision = 1e-05,
  deps = 1e-04
)
}
\arguments{
\item{deps}{}
}
\description{
Title
}
