% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_regarima}
\alias{.extended_airline_regarima}
\title{Create a RegARIMA model based on an Extended Airline specification}
\usage{
.extended_airline_regarima(series, jspec, mean = FALSE, xreg = NULL)
}
\arguments{
\item{series}{time series}

\item{jspec}{A Java ExtendedAirlineSpec object, for instance created using
\code{\link{.extended_airline_spec}}.}

\item{mean}{Logical. If \code{TRUE}, includes a mean correction term in the model.
Default: \code{FALSE}.}

\item{xreg}{Optional matrix of regression variables.
Default: \code{NULL} (no regressors).}
}
\value{
A Java object of class \code{RegArimaModel} from the JDemetra+ toolkit.
}
\description{
This internal function constructs a Java RegArimaModel object by fitting an
Extended Airline model.
}
\keyword{internal}
