% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extendedairline_estimation}
\alias{.extendedairline_estimation}
\title{Title}
\usage{
.extendedairline_estimation(
  jregarima,
  jspec,
  eps = 1e-09,
  exactHessian = FALSE
)
}
\arguments{
\item{exactHessian}{}
}
\description{
Title
}
