% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.plot_jd}
\alias{.plot_jd}
\title{Create a custom plot replicating the JD+ GUI visual style}
\usage{
.plot_jd(x, y, col, legend_txt = NULL, ...)
}
\arguments{
\item{x}{Numeric vector representing the x-axis values, typically a R time 
series object.}

\item{y}{A list of numeric vectors, where each element represents a
different time series to be plotted on the y-axis.}

\item{col}{A character or color vector specifying the color assigned to each
series.}

\item{legend_txt}{An optional character vector of labels for the legend.}

\item{...}{Additional graphical parameters. Can be used to customize line
types (\code{lty}), line widths (\code{lwd}), or other aesthetic properties
that apply uniformly to all series.}
}
\value{
`NULL` (invisible).
}
\description{
This internal function generates a time series plot that replicates the visual
template of the JD+ graphical user interface. The function 
supports plotting one or more time series simultaneously.
}
