% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.R
\name{.r2jd_doubleseq}
\alias{.r2jd_doubleseq}
\title{Convert an R numeric vector to a Java DoubleSeq object}
\usage{
.r2jd_doubleseq(x)
}
\arguments{
\item{x}{An object that can be coerced to a numeric vector.}
}
\value{
A Java object of class
\code{jdplus.toolkit.base.api.data.DoubleSeq}.
}
\description{
This internal helper function converts an R object to a Java
\code{DoubleSeq} object from the JDemetra+ toolkit.
}
\details{
The input is first coerced to numeric using \code{as.numeric()} and is then
passed to the static Java factory method \code{DoubleSeq.of()}.


This function is mainly intended for internal use when passing numeric R
vectors to Java methods that expect a \code{DoubleSeq}.
}
\examples{
\dontrun{
x <- c(1.2, 3.4, 5.6)
jseq <- .r2jd_doubleseq(x)
}

}
\keyword{internal}
