% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{.x11plus_default}
\alias{.x11plus_default}
\title{Perform an X-11 like decomposition with any (non-integer) periodicity.}
\usage{
.x11plus_default(
  series,
  period,
  multiplicative = TRUE,
  trend_horizon = 6,
  trend_degree = 3,
  trend_kernel = c("Henderson", "BiWeight", "TriWeight", "TriCube", "Uniform",
    "Triangular", "Epanechnikov", "Trapezoidal"),
  trend_asymmetric = c("CutAndNormalize", "Direct", "MMSRE"),
  seas_s0 = c("S3X3", "S3X1", "S3X5", "S3X9", "S3X15"),
  seas_s1 = c("S3X5", "S3X3", "S3X1", "S3X9", "S3X15"),
  extreme_lsig = 1.5,
  extreme_usig = 2.5,
  user_defined = NULL
)
}
\arguments{
\item{series}{input time-series.}

\item{period}{Period of the seasonal component, any positive real number.}

\item{multiplicative}{Boolean indicating if the decomposition mode is multiplicative (TRUE).}

\item{trend_horizon}{bandwidth of trend filters.}

\item{trend_degree}{polynomial order in local trend model.}

\item{trend_kernel}{kernel weights in objective function.}

\item{trend_asymmetric}{truncation type for symmetric filter.
\code{"matrix"}, \code{"lp_filter"} or \code{"rkhs_filter"}.}

\item{seas_s0}{Seasonal filter for B5, C5, D5.}

\item{seas_s1}{seasonal filter for B10, C10, D10.}

\item{extreme_lsig}{lower boundary used for outlier correction in irregular.}

\item{extreme_usig}{upper boundary used for outlier correction in irregular.}

\item{user_defined}{a vector containing the additional output variables.}
}
\value{
An object of the class 'hf_decomposition', containing the decomposition and the parameters
}
\description{
Perform an X-11 like decomposition with any (non-integer) periodicity.
}
\details{
If trend_coefs are provided, the other trend-settings are ignored and .x11plus_trend() is run.
}
\examples{
q <- x11plus(rjd3toolkit::ABS$X0.2.09.10.M, 12)
}
\keyword{internal}
