% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{.x11plus_trend}
\alias{.x11plus_trend}
\title{X-11 Decomposition With Custom Trend Filters}
\usage{
.x11plus_trend(
  series,
  period = stats::frequency(series),
  trend_coefs,
  multiplicative = TRUE,
  seas_s0 = c("S3X3", "S3X1", "S3X5", "S3X9", "S3X15"),
  seas_s1 = c("S3X5", "S3X3", "S3X1", "S3X9", "S3X15"),
  extreme_lsig = 1.5,
  extreme_usig = 2.5,
  user_defined = NULL
)
}
\arguments{
\item{series}{input time-series.}

\item{period}{period.}

\item{trend_coefs}{coefficients of the filters used for the trend-cycle extraction from
the real-time asymmetric filter to the symmetric filter. Can be a, object of class \code{"list"},
\code{"matrix"}, \code{"lp_filter"} or \code{"rkhs_filter"}.}

\item{multiplicative}{boolean indicating if the decomposition mode is multiplicative.}

\item{seas_s0, seas_s1}{seasonal filters.}

\item{extreme_lsig, extreme_usig}{boundaries used for outlier correction in irregular.}

\item{user_defined}{a vector containing the additional output variables.}
}
\description{
Perform the X-11 decomposition using custom trend filter
}
\keyword{internal}
