% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x11correction.R
\name{extreme_values_correction}
\alias{extreme_values_correction}
\title{'X11' Extreme Values Corrector}
\usage{
extreme_values_correction(
  series,
  period,
  corrected_s,
  lsigma = 1.5,
  usigma = 2.5,
  multiplicative = FALSE,
  start = 0,
  clean_extremities = TRUE
)
}
\arguments{
\item{series}{the analysed time series.}

\item{period}{the period of the input time series if `series` is not a `"ts"` object.}

\item{corrected_s}{other time series if the series being corrected is different from series.}

\item{lsigma}{the lower sigma boundaries for the detection of extreme values.}

\item{usigma}{the upper sigma boundaries for the detection of extreme values.}

\item{multiplicative}{boolean indicating if the decomposition is multiplicative or additive.}

\item{start}{position of the first "complete" considered period.}

\item{clean_extremities}{boolean indicating if the extremities should be cleaned.}
}
\description{
'X11' Extreme Values Corrector
}
\details{
The 'X11' Extreme Values Corrector is used to compute the tables
b4, b4g, b9, b9g, b17, b20, c17 and c20.

#' The returned correction contains the following columns:

\describe{
  \item{\code{obs_weight}}{Extreme value corrections weight of each observation}
  \item{\code{correction_factors}}{Extreme value correction factors}
}
}
\examples{
extreme_values_correction(rjd3toolkit::ABS$X0.2.09.10.M, 12)

}
