% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition.ssf}
\alias{fractionalAirlineDecomposition.ssf}
\title{Title}
\usage{
fractionalAirlineDecomposition.ssf(jdecomp)
}
\arguments{
\item{jdecomp}{}
}
\value{

}
\description{
Title
}
