% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{henderson}
\alias{henderson}
\title{Apply Henderson linear filter}
\usage{
henderson(series, length, musgrave = TRUE, ic = 4.5)
}
\arguments{
\item{series}{Input time series.}

\item{length}{Length of the symmetric Henderson filter. Must be odd.}

\item{musgrave}{Boolean indicating if Musgrave asymmetric filters should be used. Default is \code{TRUE}. If \code{FALSE}, no asymmetric filters will be used.}

\item{ic}{I/C ratio between the volatility of the tentative irregular and trend-cycle estimates (needed for calculating the Musgrave asymmetric filters).}
}
\value{
A numeric array corresponding to the estimated trend.
}
\description{
Apply Henderson linear filter
}
\examples{
q <- x11plus(rjd3toolkit::ABS$X0.2.09.10.M, 12)

henderson(q$decomposition$sa, 13)
}
