% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{Holidays}
\alias{Holidays}
\title{Constructor for a holidays calendar}
\usage{
Holidays(...)
}
\arguments{
\item{...}{a serie of holiday items. One of :
* a prespecified calender, see [listPredefinedCalendars()]
* a prespecified holiday, see [listPrespecifiedHolidays()]
* an integer vector of length 2 (month, day) indication a specific day
* a named list containing as first item one of the above possibilities (excluding a calender) with in addition optional the following named elementd :
   * _offset_ : to indicate multiple days around specified day
   * _weight_ : optional a weight
   * _gregorian_  : a logical}
}
\value{
[Holidays] a named list containing the calendar created in JDemetra. In addition it contains some labels and the used arguments to construct it.
}
\description{
Constructor for a holidays calendar
}
\examples{
\dontrun{
  # Just one day on 12th june
  Holidays(c(6,12))
  # Just one day being a prespecified day called _Christmas°
  Holidays("Christmas")
  # combining both
  Holidays(c(6,12), "Christmas")
  # combining both but also the day before and after it included
  Holidays(list(c(6,12), offset=-1:1),
           list("Christmas", offset=-1:1))
  # Reuse default calendar already specified for Belgium
  Holidays("Belgium")
  # Reuse default calendar already specified for Belgium but add some more days
  Holidays("Belgium", c(6, 12))
}
}
