% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{HolidaysMatrix}
\alias{HolidaysMatrix}
\title{Constructor for a holiday matrix in JDemetra, with time as rows and holidays as columns. A flag indicates if it is an holiday or not.}
\usage{
HolidaysMatrix(
  holidays,
  startingDate,
  length,
  nonworking = as.integer(7),
  type = c("Default", "Skip", "NextWorkingDay", "PreviousWorkingDay")
)
}
\arguments{
\item{startingDate}{a [Date] or a character to coerce to a Date being the first day as from which the matrix is made}

\item{length}{[integer] the number of days to include}

\item{nonworking}{[integer] An array with the non working days (Mondays=1,...,Saturdays=6, Sundays=7). Usually 7, or c(6,7)}

\item{type}{[character] specific treatment of non working days around working days ... See JDemetra.}

\item{a}{holidays calender specification constructed by [Holidays()]}
}
\value{
JDMatrix which is a named list containing the regression matrix created in JDemetra. In addition it contains the used arguments to construct it.
}
\description{
Constructor for a holiday matrix in JDemetra, with time as rows and holidays as columns. A flag indicates if it is an holiday or not.
}
