% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_r.R
\name{ic_ratio}
\alias{ic_ratio}
\title{Compute IC-Ratio}
\usage{
ic_ratio(
  series,
  frequency = NULL,
  trend_length = NULL,
  musgrave = FALSE,
  multiplicative = FALSE,
  seasonally_adjust = TRUE,
  sma = "S3X5",
  ...
)
}
\arguments{
\item{series}{input time series}

\item{frequency}{frequency of the seasonal component. Can be NULL if series is ts}

\item{trend_length}{length of Henderson filter used for trend adjustment. If NULL it is nextodd(frequency)}

\item{musgrave}{Boolean indicating if Musgrave asymmetric filters should be used in trend adjustment}

\item{multiplicative}{boolean indicating if the decomposition is multiplicative or additive}

\item{seasonally_adjust}{boolean. Should series be seasonally adjusted?}

\item{sma}{seasonal moving average used in x11plus()}

\item{...}{additional arguments for trend adjustment. See ?henderson}
}
\description{
Compute IC-Ratio
}
\details{
In the traditional calculation of the IC-ratio, no asymmetric filters used for the preliminary
trend estimation (see Ladiray and Quenneville 2001, Table B7).
}
\examples{
series <- rjd3toolkit::Retail$AllOtherGenMerchandiseStores
ic_ratio(series)

}
\references{
Ladiray, D., Quenneville, B. (2001). The Various Tables. In: Seasonal Adjustment with the X-11 Method. Lecture Notes in Statistics, vol 158. Springer, New York, NY. https://doi.org/10.1007/978-1-4613-0175-2_5
}
