% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{localpolynomials}
\alias{localpolynomials}
\title{Apply local polynomials filters.}
\usage{
localpolynomials(
  y,
  horizon,
  degree = 3,
  kernel = c("Henderson", "Uniform", "Biweight", "Triweight", "Tricube", "Gaussian",
    "Triangular", "Parabolic"),
  endpoints = c("DAF", "CC", "LC", "QL", "CQ"),
  ic = 4.5
)
}
\arguments{
\item{y}{input time series.}

\item{horizon}{horizon (bandwidth) of the symmetric filter.}

\item{degree}{degree degree of polynomial.}

\item{kernel}{type of kernel weights used.}

\item{endpoints}{method for endpoints.}

\item{ic}{ic ratio: irregular/trend-cycle.}
}
\description{
Apply local polynomials filters.
}
