% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{loglevel}
\alias{loglevel}
\title{Log-Level test for Extended Airline Models}
\usage{
loglevel(input, precision = 1e-05, deps = 1e-04)
}
\arguments{
\item{input}{time series}

\item{precision}{Numeric value specifying the tolerance for convergence
of optimization algorithms. Default: 1e-5.}

\item{deps}{Numeric scalar. Step in the computation of the numerical derivatives,
used in the optimisation routine. Default:1e-4.}
}
\value{
An object containing the log-level test results. First the AICc of the model in levels and then the AICc of the model with logs
}
\description{
This function performs a statistical test to determine whether data
should be transformed to logarithmic scale or kept in level (original scale)
}
\examples{
series <- rjd3toolkit::Retail$BookStores

loglevel(series)
}
