% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hf_decomposition}
\alias{plot.hf_decomposition}
\title{Plot method for hf_decomposition objects}
\usage{
\method{plot}{hf_decomposition}(
  x,
  ...,
  col_bg = "white",
  col_grid = "#777777",
  col_labels = "#777777",
  col_1 = "#debc12",
  col_2 = "#74b6d6",
  col_3 = "#ee2c2c"
)
}
\arguments{
\item{x}{An object of class \code{hf_decomposition}.}

\item{...}{options for par()}

\item{col_bg}{Background color.}

\item{col_grid}{Color of grid lines.}

\item{col_labels}{Color of tick labels.}

\item{col_1}{First time series color.}

\item{col_2}{Second time series color.}

\item{col_3}{Third time series color.}
}
\description{
Plot method for hf_decomposition objects
}
\keyword{internal}
