% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hf_estimation}
\alias{plot.hf_estimation}
\title{Plot method for hf_estimation objects}
\usage{
\method{plot}{hf_estimation}(
  x,
  ...,
  col_bg = "#ffffe0",
  col_grid = "#777777",
  col_labels = "#777777",
  col_1 = "#debc12",
  col_2 = "#74b6d6"
)
}
\arguments{
\item{x}{An object of class \code{JDFractionalAirlineDecomposition}.}

\item{...}{additional options for first line.}

\item{col_bg}{Background color.}

\item{col_grid}{Color of grid lines.}

\item{col_labels}{Color of tick labels.}

\item{col_1}{First time series color.}

\item{col_2}{Second time series color.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot method for hf_estimation objects
}
\keyword{internal}
