% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_jd}
\alias{plot_jd}
\title{Custom Plot Function on JD+ template}
\usage{
plot_jd(x, y, col, legend_txt = NULL, ...)
}
\arguments{
\item{x}{Numeric vector, x-axis values.}

\item{y}{List of numeric vectors, y-axis values for different series.}

\item{col}{Vector of colors for different series.}

\item{legend_txt}{Character vector of legend labels for different series.}

\item{...}{Additional graphical parameters.}
}
\value{
`NULL` (invisible).
}
\description{
This function creates a customized plot in the same template as JD+ GUI color and forms.
}
