% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hf_decomposition}
\alias{print.hf_decomposition}
\title{Print method for hf_decomposition objects}
\usage{
\method{print}{hf_decomposition}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{hf_decomposition}.}

\item{digits}{Number of significant digits for numeric output.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for hf_decomposition objects
}
\keyword{internal}
