% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hf_estimation}
\alias{print.hf_estimation}
\title{Print method for hf_estimation objects}
\usage{
\method{print}{hf_estimation}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{hf_estimation}.}

\item{digits}{Number of significant digits for numeric output.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for hf_estimation objects
}
\keyword{internal}
