% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stl.R
\name{stl}
\alias{stl}
\title{Perform an STL like (based on Loess) decomposition on any periodicity}
\usage{
stl(y, period, multiplicative = TRUE, swindow = 7, twindow = 0, robust = TRUE)
}
\arguments{
\item{y}{input time series.}

\item{period}{period, any positive real number.}

\item{multiplicative}{Boolean indicating if the decomposition mode is multiplicative (TRUE).}

\item{swindow}{length of seasonal filter.}

\item{twindow}{length of trend filter.}

\item{robust}{Boolean: in outer loop robust weights for irregular.}
}
\description{
Perform an STL like (based on Loess) decomposition on any periodicity
}
