% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hf_decomposition}
\alias{summary.hf_decomposition}
\title{Summary method for hf_decomposition objects}
\usage{
\method{summary}{hf_decomposition}(object, digits = 3L, ...)
}
\arguments{
\item{object}{An object of class \code{hf_decomposition}.}

\item{digits}{Number of significant digits for numeric output.}

\item{...}{Ignored.}
}
\value{
\code{object}, invisibly.
}
\description{
Summary method for hf_decomposition objects
}
\keyword{internal}
