% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.hf_estimation}
\alias{summary.hf_estimation}
\title{Summary method for JDFractionalAirlineEstimation objects}
\usage{
\method{summary}{hf_estimation}(object, digits = 3L, ...)
}
\arguments{
\item{object}{An object of class \code{JDFractionalAirlineEstimation}.}

\item{digits}{Number of significant digits for numeric output.}

\item{...}{Ignored.}
}
\value{
\code{object}, invisibly.
}
\description{
Summary method for JDFractionalAirlineEstimation objects
}
\keyword{internal}
