% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_loglevel}
\alias{.extended_airline_loglevel}
\title{Title}
\usage{
.extended_airline_loglevel(jregarima, jspec, precision = 1e-05)
}
\arguments{
\item{precision}{}
}
\description{
Title
}
