% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_regarima}
\alias{.extended_airline_regarima}
\title{Creates a java RegArima models based on an extended airline spec}
\usage{
.extended_airline_regarima(y, jspec, mean = FALSE, X = NULL)
}
\arguments{
\item{y}{y}

\item{jspec}{Java spec}

\item{mean}{Mean correction (to be avoided)}

\item{X}{Regression variables}
}
\value{
A Java RegArima model
}
\description{
Creates a java RegArima models based on an extended airline spec
}
\examples{
jspec<-.extended_airline_spec(c(12))
.extended_airline_regarima(rjd3toolkit::ABS$X0.2.09.10.M, jspec)
}
