% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_extendedairline.R
\name{.extended_airline_spec}
\alias{.extended_airline_spec}
\title{Internal routine to create an ExtendedAirlineSpec}
\usage{
.extended_airline_spec(
  periodicities,
  differencing = -1,
  ar = FALSE,
  toint = FALSE
)
}
\arguments{
\item{periodicities}{Periodicities}

\item{differencing}{Differnecing order. -1 for automatic computation}

\item{ar}{Use of an AR regular stationary polynomial instead of a MA polynomial}

\item{toint}{Round periodicties to integers}
}
\value{
A Java ExtendedAirlineSpec object
}
\description{
Internal routine to create an ExtendedAirlineSpec
}
\examples{
.extended_airline_spec(c(7, 365.25))
}
