% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition}
\alias{fractionalAirlineDecomposition}
\title{Perform an Arima Model Based (AMB) decomposition}
\usage{
fractionalAirlineDecomposition(
  y,
  period,
  sn = F,
  stde = F,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{y}{input time series}

\item{period}{period of the seasonal component, any positive real number.}

\item{sn}{decomposition into signal and noise. The signal is the seasonally adjusted series and the noise the seasonal component.}

\item{stde}{Boolean: TRUE if standard deviations of the components must be computed. In some cases (memory limits), it is currently not possible to compute them}

\item{nbcasts}{number of backcasts.}

\item{nfcasts}{number of forecasts.}

\item{adjust}{Boolean: TRUE if an actual fractional airline model is used, FALSE if the period is rounded to the nearest integer.}
}
\description{
Perform an Arima Model Based (AMB) decomposition
}
