% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition_raw}
\alias{fractionalAirlineDecomposition_raw}
\title{Title}
\usage{
fractionalAirlineDecomposition_raw(
  y,
  period,
  sn = F,
  stde = F,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{y}{}

\item{period}{}

\item{sn}{}

\item{stde}{}

\item{nbcasts}{}

\item{nfcasts}{}
}
\description{
Title
}
