% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition_raw}
\alias{fractionalAirlineDecomposition_raw}
\title{Title}
\usage{
fractionalAirlineDecomposition_raw(
  y,
  period,
  sn = FALSE,
  stde = FALSE,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{nfcasts}{}
}
\description{
Title
}
