% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineDecomposition_ssf}
\alias{fractionalAirlineDecomposition_ssf}
\title{Title}
\usage{
fractionalAirlineDecomposition_ssf(jdecomp)
}
\arguments{
\item{jdecomp}{}
}
\description{
Title
}
