% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineEstimation}
\alias{fractionalAirlineEstimation}
\title{Linearize the series with a fractional airline model}
\usage{
fractionalAirlineEstimation(
  y,
  periods,
  x = NULL,
  ndiff = 2,
  ar = F,
  mean = FALSE,
  outliers = NULL,
  criticalValue = 6,
  precision = 1e-12,
  approximateHessian = F
)
}
\arguments{
\item{y}{input time series.}

\item{periods}{vector of periods values of the seasonal component, any positive real numbers.}

\item{x}{matrix of user-defined regression variables (see rjd3modelling for building calendar regressors).}

\item{mean}{add constant mean to y after differencing.}

\item{outliers}{type of outliers sub vector of c("AO","LS","WO")}

\item{criticalValue}{critical value for automatic outlier detection}
}
\description{
Linearize the series with a fractional airline model
}
