% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{fractionalAirlineEstimation}
\alias{fractionalAirlineEstimation}
\title{Linearize the series with a fractional airline model}
\usage{
fractionalAirlineEstimation(
  y,
  periods,
  x = NULL,
  ndiff = 2,
  ar = FALSE,
  outliers = NULL,
  criticalValue = 6,
  precision = 1e-12,
  approximateHessian = FALSE,
  nfcasts = 0,
  log = FALSE,
  y_time = NULL
)
}
\arguments{
\item{y}{input time series.}

\item{periods}{vector of periods values of the seasonal component, any positive real numbers.}

\item{x}{matrix of user-defined regression variables (see rjd3toolkit for building calendar regressors).}

\item{outliers}{type of outliers sub vector of c("AO","LS","WO")}

\item{criticalValue}{Critical value for automatic outlier detection}

\item{precision}{Precision of the likelihood}

\item{approximateHessian}{Compute approximate hessian (based on the optimizing procedure)}

\item{nfcasts}{Number of forecasts}

\item{log}{a logical}

\item{y_time}{vector of times at which `y` is indexed}
}
\description{
Linearize the series with a fractional airline model
}
