% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{jd2r_fractionalAirlineDecomposition}
\alias{jd2r_fractionalAirlineDecomposition}
\title{Title}
\usage{
jd2r_fractionalAirlineDecomposition(
  jrslt,
  sn = FALSE,
  stde = FALSE,
  period,
  log = FALSE,
  y_time = NULL
)
}
\arguments{
\item{y_time}{vector of times at which the time series is indexed}
}
\description{
Title
}
