% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{jd2r_fractionalAirlineDecomposition}
\alias{jd2r_fractionalAirlineDecomposition}
\title{Title}
\usage{
jd2r_fractionalAirlineDecomposition(jrslt, sn = F, stde = F, period)
}
\arguments{
\item{jrslt}{}

\item{sn}{}

\item{stde}{}
}
\description{
Title
}
