% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{jd2r_multiAirlineDecomposition}
\alias{jd2r_multiAirlineDecomposition}
\title{Title}
\usage{
jd2r_multiAirlineDecomposition(
  jrslt,
  stde = F,
  periods,
  log = FALSE,
  y_time = NULL
)
}
\arguments{
\item{jrslt}{}

\item{stde}{}

\item{y_time}{vector of times at which the time series is indexed}
}
\description{
Title
}
