% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stl.R
\name{loess}
\alias{loess}
\title{Fit a Loess regression.}
\usage{
loess(y, window, degree = 1, jump = 1)
}
\arguments{
\item{y}{input time series.}

\item{window}{}

\item{degree}{}

\item{jump}{}
}
\description{
Fit a Loess regression.
}
