% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{multiAirlineDecomposition}
\alias{multiAirlineDecomposition}
\title{Perform an Arima Model Based (AMB) decomposition on several periodcities at once}
\usage{
multiAirlineDecomposition(
  y,
  periods,
  ndiff = 2,
  ar = F,
  stde = F,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{y}{input time series.}

\item{periods}{vector of periods values of the seasonal component, any positive real numbers.}

\item{stde}{Boolean: TRUE: compute standard deviations of the components. In some cases (memory limits), it is currently not possible to compute them}

\item{nbcasts}{number of backcasts.}

\item{nfcasts}{number of forecasts.}

\item{adjust}{Boolean: TRUE: actual fractional airline model is to be used, FALSE: the period is rounded to the nearest integer.}

\item{sn}{decomposition into signal and noise (2 components only). The signal is the seasonally adjusted series and the noise the seasonal component.}
}
\description{
Perform an Arima Model Based (AMB) decomposition on several periodcities at once
}
