% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_fractionalairline.R
\name{multiAirlineDecomposition_raw}
\alias{multiAirlineDecomposition_raw}
\title{Title}
\usage{
multiAirlineDecomposition_raw(
  y,
  periods,
  ndiff = 2,
  ar = FALSE,
  stde = FALSE,
  nbcasts = 0,
  nfcasts = 0
)
}
\arguments{
\item{nfcasts}{}
}
\description{
Title
}
