% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.JDFractionalAirlineDecomposition}
\alias{plot.JDFractionalAirlineDecomposition}
\title{Plot Function for JDFractionalAirlineDecomposition Objects}
\usage{
\method{plot}{JDFractionalAirlineDecomposition}(x, from, to, type_chart = c("y-sa-trend", "cal-seas-irr"), ...)
}
\arguments{
\item{x}{An object of class 'JDFractionalAirlineDecomposition'.}

\item{from}{`Date` or `POSIXt` object, optional starting point for x-axis.}

\item{to}{`Date` or `POSIXt` object, optional ending point for x-axis.}

\item{type_chart}{Character vector specifying the type of chart to plot ("y-sa-trend", "cal-seas-irr").}

\item{...}{Additional graphical parameters.}
}
\value{
`NULL` (invisible).
}
\description{
This function creates a plot for the result of an Arima Model Based (AMB) decomposition of one or several frequencies (class `JDFractionalAirlineDecomposition`). It shows the decomposition and the component of the model.
}
