% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.JDFractionalAirlineEstimation}
\alias{plot.JDFractionalAirlineEstimation}
\title{Plot Function for JDFractionalAirlineEstimation Objects}
\usage{
\method{plot}{JDFractionalAirlineEstimation}(x, from, to, ...)
}
\arguments{
\item{x}{An object of class 'JDFractionalAirlineEstimation'.}

\item{from}{`Date` or `POSIXt` object, optional starting point for x-axis.}

\item{to}{`Date` or `POSIXt` object, optional ending point for x-axis.}

\item{...}{Additional graphical parameters.}
}
\value{
`NULL` (invisible).
}
\description{
This function creates a plot for the result of fractional airline model (class `JDFractionalAirlineEstimation`). It shows the raw data and linearized series.
}
