% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.JDFractionalAirlineEstimation}
\alias{print.JDFractionalAirlineEstimation}
\title{Print method for 'JDFractionalAirlineEstimation' objects}
\usage{
\method{print}{JDFractionalAirlineEstimation}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class 'JDFractionalAirlineEstimation'.}

\item{digits}{Number of digits to round numerical values (default is 3 or digits - 3 from options).}
}
\value{
The original object 'x'.
}
\description{
This function prints informations on the result of a Fractional Airline model (classe JDFractionalAirlineEstimation).
}
